split useWorldUsers into it's own file
