+++ src/hooks/users/index.ts
@@ -5,7 +5,7 @@ import { User } from "types/User";
 import { WithId } from "utils/id";
 import { normalizeTimestampToMilliseconds } from "utils/time";
 
-import { worldUsersByIdSelector, worldUsersSelector } from "utils/selectors";
+import { worldUsersByIdSelector } from "utils/selectors";
 
 import { useConnectCurrentVenueNG } from "hooks/useConnectCurrentVenueNG";
 import { isLoaded } from "hooks/useFirestoreConnect";
@@ -14,25 +14,10 @@ import { useSelector } from "hooks/useSelector";
 import { useVenueId } from "hooks/useVenueId";
 
 import { useConnectWorldUsers } from "./useConnectWorldUsers";
+import { useWorldUsers } from "./useWorldUsers";
 
 export { useConnectWorldUsers } from "./useConnectWorldUsers";
-
-export const useWorldUsers = (): {
-  worldUsers: readonly WithId<User>[];
-  isWorldUsersLoaded: boolean;
-} => {
-  useConnectWorldUsers();
-
-  const selectedWorldUsers = useSelector(worldUsersSelector);
-
-  return useMemo(
-    () => ({
-      worldUsers: selectedWorldUsers ?? [],
-      isWorldUsersLoaded: isLoaded(selectedWorldUsers),
-    }),
-    [selectedWorldUsers]
-  );
-};
+export { useWorldUsers } from "./useWorldUsers";
 
 // @debt typing, Record implies that a User will exist for literally any given string, which is untrue
 export const useWorldUsersById = () => {
